// web/nestspend-web/src/app/core/api/models/register-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Request body for user registration
 */
export interface RegisterRequest {

  /**
   * User's display name
   */
  displayName: string;

  /**
   * User's email address
   */
  email: string;

  /**
   * Name of the household to create
   */
  householdName: string;

  /**
   * User's password (8-100 characters)
   */
  password: string;
}